This should add the final permission checks

+++ src/resolvers/judge-assignment.ts
@@ -5,14 +5,16 @@ import { JudgeAssignmentDoc } from '../store/schema'
 
 export const judgeAssignmentResolvers: Resolvers = {
   Mutation: {
-    async createJudgeAssignment (_, { judgeId, categoryId, data }, { dataSources, allowUser }) {
-      const judge = await dataSources.judges.findOneById(judgeId, { ttl: Ttl.Short })
+    async createJudgeAssignment (_, { judgeId, categoryId, data }, { dataSources, allowUser, user }) {
+      const judge = await dataSources.judges.findOneById(judgeId)
       if (!judge) throw new ApolloError('Judge does not exist')
-      const category = await dataSources.categories.findOneById(categoryId, { ttl: Ttl.Short })
+      const category = await dataSources.categories.findOneById(categoryId)
       if (!category) throw new ApolloError('Category does not exist')
 
-      const group = await dataSources.groups.findOneById(judge.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).judge(judge).assign.assert()
+      const group = await dataSources.groups.findOneById(judge.groupId)
+      if (!group) throw new ApolloError('Group does not exist')
+      const authJudge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id })
+      allowUser.group(group, authJudge).category(category).judge(judge).assign.assert()
 
       await dataSources.judgeAssignments.createOne({
         judgeId: judge.id,
@@ -26,22 +28,24 @@ export const judgeAssignmentResolvers: Resolvers = {
 
       return judge
     },
-    async updateJudgeAssignment (_, { judgeAssignmentId, data }, { dataSources, allowUser }) {
+    async updateJudgeAssignment (_, { judgeAssignmentId, data }, { dataSources, allowUser, user }) {
       const assignment = await dataSources.judgeAssignments.findOneById(judgeAssignmentId)
       if (!assignment) throw new ApolloError('Assignment does not exist')
-      const judge = await dataSources.judges.findOneById(assignment.judgeId, { ttl: Ttl.Short })
+      const judge = await dataSources.judges.findOneById(assignment.judgeId)
       if (!judge) throw new ApolloError('Judge does not exist')
-      const category = await dataSources.categories.findOneById(assignment.categoryId, { ttl: Ttl.Short })
+      const category = await dataSources.categories.findOneById(assignment.categoryId)
       if (!category) throw new ApolloError('Category does not exist')
 
-      const group = await dataSources.groups.findOneById(judge.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).judge(judge).assign.assert()
+      const group = await dataSources.groups.findOneById(judge.groupId)
+      if (!group) throw new ApolloError('Group does not exist')
+      const authJudge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id })
+      allowUser.group(group, authJudge).category(category).judge(judge).assign.assert()
 
       return await dataSources.judgeAssignments.updateOnePartial(assignment.id, {
         options: data.options
       }) as JudgeAssignmentDoc
     },
-    async deleteJudgeAssignment (_, { judgeAssignmentId }, { dataSources, allowUser }) {
+    async deleteJudgeAssignment (_, { judgeAssignmentId }, { dataSources, allowUser, user }) {
       const assignment = await dataSources.judgeAssignments.findOneById(judgeAssignmentId)
       if (!assignment) throw new ApolloError('Assignment does not exist')
       const judge = await dataSources.judges.findOneById(assignment.judgeId, { ttl: Ttl.Short })
@@ -50,7 +54,9 @@ export const judgeAssignmentResolvers: Resolvers = {
       if (!category) throw new ApolloError('Category does not exist')
 
       const group = await dataSources.groups.findOneById(judge.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).judge(judge).assign.assert()
+      if (!group) throw new ApolloError('Group does not exist')
+      const authJudge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id })
+      allowUser.group(group, authJudge).category(category).judge(judge).assign.assert()
 
       await dataSources.judgeAssignments.deleteOne(assignment.id)
 
@@ -58,19 +64,22 @@ export const judgeAssignmentResolvers: Resolvers = {
     }
   },
   JudgeAssignment: {
-    async judge (judgeAssignment, args, { dataSources, allowUser }) {
+    async judge (judgeAssignment, args, { dataSources, allowUser, user }) {
       const judge = await dataSources.judges.findOneById(judgeAssignment.judgeId, { ttl: Ttl.Short })
       if (!judge) throw new ApolloError('Judge does not exist')
       const group = await dataSources.groups.findOneById(judge.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).getUsers.assert()
+      const authJudge = await dataSources.judges.findOneByActor({ actor: user, groupId: judge.groupId }, { ttl: Ttl.Short })
+      allowUser.group(group, authJudge).getUsers.assert()
 
       return judge
     },
-    async category (judgeAssignment, args, { dataSources, allowUser }) {
+    async category (judgeAssignment, args, { dataSources, allowUser, user }) {
       const category = await dataSources.categories.findOneById(judgeAssignment.categoryId, { ttl: Ttl.Short })
       if (!category) throw new ApolloError('Category does not exist')
       const group = await dataSources.groups.findOneById(category.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).get.assert()
+      if (!group) throw new ApolloError('Group does not exist')
+      const authJudge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id }, { ttl: Ttl.Short })
+      allowUser.group(group, authJudge).category(category).get.assert()
 
       return category
     }
