+++ src/resolvers/participant.ts
@@ -5,11 +5,13 @@ import { AthleteDoc, isAthlete, isTeam, TeamDoc } from '../store/schema'
 
 export const participantResolvers: Resolvers = {
   Mutation: {
-    async createAthlete (_, { categoryId, data }, { allowUser, dataSources }) {
-      const category = await dataSources.categories.findOneById(categoryId, { ttl: Ttl.Short })
+    async createAthlete (_, { categoryId, data }, { allowUser, dataSources, user }) {
+      const category = await dataSources.categories.findOneById(categoryId)
       if (!category) throw new ApolloError('Category does not exist')
-      const group = await dataSources.groups.findOneById(category.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).update.assert()
+      const group = await dataSources.groups.findOneById(category.groupId)
+      if (!group) throw new ApolloError('Group does not exist')
+      const judge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id })
+      allowUser.group(group, judge).category(category).update.assert()
 
       await dataSources.participants.createOne({
         categoryId,
@@ -22,11 +24,13 @@ export const participantResolvers: Resolvers = {
 
       return category
     },
-    async createTeam (_, { categoryId, data }, { allowUser, dataSources }) {
-      const category = await dataSources.categories.findOneById(categoryId, { ttl: Ttl.Short })
+    async createTeam (_, { categoryId, data }, { allowUser, dataSources, user }) {
+      const category = await dataSources.categories.findOneById(categoryId)
       if (!category) throw new ApolloError('Category does not exist')
-      const group = await dataSources.groups.findOneById(category.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).update.assert()
+      const group = await dataSources.groups.findOneById(category.groupId)
+      if (!group) throw new ApolloError('Group does not exist')
+      const judge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id })
+      allowUser.group(group, judge).category(category).update.assert()
 
       await dataSources.participants.createOne({
         categoryId,
@@ -40,13 +44,15 @@ export const participantResolvers: Resolvers = {
       return category
     },
 
-    async updateAthlete (_, { participantId, data }, { dataSources, allowUser }) {
-      const participant = await dataSources.participants.findOneById(participantId, { ttl: Ttl.Short })
+    async updateAthlete (_, { participantId, data }, { dataSources, allowUser, user }) {
+      const participant = await dataSources.participants.findOneById(participantId)
       if (!participant) throw new ApolloError('Participant does not exist')
-      const category = await dataSources.categories.findOneById(participant.categoryId, { ttl: Ttl.Short })
+      const category = await dataSources.categories.findOneById(participant.categoryId)
       if (!category) throw new ApolloError('Category does not exist')
-      const group = await dataSources.groups.findOneById(category.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).update.assert()
+      const group = await dataSources.groups.findOneById(category.groupId)
+      if (!group) throw new ApolloError('Group does not exist')
+      const judge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id })
+      allowUser.group(group, judge).category(category).update.assert()
 
       if (!isAthlete(participant)) throw new ApolloError('Participant is not an athlete')
 
@@ -59,13 +65,15 @@ export const participantResolvers: Resolvers = {
 
       return await dataSources.participants.updateOnePartial(participant.id, updates) as AthleteDoc
     },
-    async updateTeam (_, { participantId, data }, { dataSources, allowUser }) {
-      const participant = await dataSources.participants.findOneById(participantId, { ttl: Ttl.Short })
+    async updateTeam (_, { participantId, data }, { dataSources, allowUser, user }) {
+      const participant = await dataSources.participants.findOneById(participantId)
       if (!participant) throw new ApolloError('Participant does not exist')
-      const category = await dataSources.categories.findOneById(participant.categoryId, { ttl: Ttl.Short })
+      const category = await dataSources.categories.findOneById(participant.categoryId)
       if (!category) throw new ApolloError('Category does not exist')
-      const group = await dataSources.groups.findOneById(category.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).update.assert()
+      const group = await dataSources.groups.findOneById(category.groupId)
+      if (!group) throw new ApolloError('Group does not exist')
+      const judge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id })
+      allowUser.group(group, judge).category(category).update.assert()
 
       if (!isTeam(participant)) throw new ApolloError('Participant is not a team')
 
@@ -79,13 +87,15 @@ export const participantResolvers: Resolvers = {
       return await dataSources.participants.updateOnePartial(participant.id, updates) as TeamDoc
     },
 
-    async deleteParticipant (_, { participantId }, { dataSources, allowUser }) {
-      const participant = await dataSources.participants.findOneById(participantId, { ttl: Ttl.Short })
+    async deleteParticipant (_, { participantId }, { dataSources, allowUser, user }) {
+      const participant = await dataSources.participants.findOneById(participantId)
       if (!participant) throw new ApolloError('Participant does not exist')
-      const category = await dataSources.categories.findOneById(participant.categoryId, { ttl: Ttl.Short })
+      const category = await dataSources.categories.findOneById(participant.categoryId)
       if (!category) throw new ApolloError('Category does not exist')
-      const group = await dataSources.groups.findOneById(category.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).update.assert()
+      const group = await dataSources.groups.findOneById(category.groupId)
+      if (!group) throw new ApolloError('Group does not exist')
+      const judge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id })
+      allowUser.group(group, judge).category(category).update.assert()
 
       await dataSources.participants.deleteOne(participant.id)
 
@@ -93,21 +103,25 @@ export const participantResolvers: Resolvers = {
     }
   },
   Athlete: {
-    async category (athlete, args, { dataSources, allowUser }) {
+    async category (athlete, args, { dataSources, allowUser, user }) {
       const category = await dataSources.categories.findOneById(athlete.categoryId, { ttl: Ttl.Short })
       if (!category) throw new ApolloError('Category does not exist')
       const group = await dataSources.groups.findOneById(category.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).get.assert()
+      if (!group) throw new ApolloError('Group does not exist')
+      const judge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id }, { ttl: Ttl.Short })
+      allowUser.group(group, judge).category(category).get.assert()
 
       return category
     }
   },
   Team: {
-    async category (athlete, args, { dataSources, allowUser }) {
+    async category (athlete, args, { dataSources, allowUser, user }) {
       const category = await dataSources.categories.findOneById(athlete.categoryId, { ttl: Ttl.Short })
       if (!category) throw new ApolloError('Category does not exist')
       const group = await dataSources.groups.findOneById(category.groupId, { ttl: Ttl.Short })
-      allowUser.group(group).category(category).get.assert()
+      if (!group) throw new ApolloError('Group does not exist')
+      const judge = await dataSources.judges.findOneByActor({ actor: user, groupId: group.id }, { ttl: Ttl.Short })
+      allowUser.group(group, judge).category(category).get.assert()
 
       return category
     }
